# api/internal/httpapi/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Let It Call Lead Generation API",
    "version": "1.0.0",
    "description": "A compatibility API for reading users, event types, availability, scheduled events, invitees, and installing webhooks. All instants use UTC RFC3339 second precision. Webhook delivery is at least once and unordered. Receivers should deduplicate with <event>:<resource URI>:<updated_at>. Signed deliveries use Webhook-Signature: t=<unix>,v1=<HMAC-SHA256(timestamp + '.' + raw body)> and receivers should enforce a 180 second replay tolerance. Routing form subscriptions are accepted, but routing form events are not emitted until routing forms are implemented."
  },
  "servers": [{ "url": "__BASE_URL__" }],
  "security": [{ "bearerAuth": [] }],
...
  "paths": {
...
    "/event_types": {
      "get": {
        "tags": ["Event Types"],
        "operationId": "listEventTypes",
        "parameters": [
          { "name": "user", "in": "query", "schema": { "type": "string", "format": "uri" } },
          { "name": "organization", "in": "query", "schema": { "type": "string", "format": "uri" } },
          { "name": "active", "in": "query", "schema": { "type": "boolean" } },
          { "name": "admin_managed", "in": "query", "schema": { "type": "boolean" } },
          { "name": "sort", "in": "query", "schema": { "type": "string", "enum": ["name", "name:asc", "name:desc"] } },
          { "$ref": "#/components/parameters/Count" },
          { "$ref": "#/components/parameters/PageToken" }
        ],
        "responses": {
          "200": { "description": "Event types", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/EventTypeCollection" } } } },
          "400": { "$ref": "#/components/responses/InvalidArgument" },
          "401": { "$ref": "#/components/responses/Unauthenticated" },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      }
    },
    "/event_types/{uuid}": {
      "get": {
        "tags": ["Event Types"],
        "operationId": "getEventType",
        "parameters": [{ "name": "uuid", "in": "path", "required": true, "schema": { "type": "string" } }],
        "responses": {
          "200": { "description": "Event type", "content": { "application/json": { "schema": { "type": "object", "required": ["resource"], "properties": { "resource": { "$ref": "#/components/schemas/EventType" } } } } } },
          "401": { "$ref": "#/components/responses/Unauthenticated" },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": { "bearerAuth": { "type": "http", "scheme": "bearer", "bearerFormat": "personal access token" } },
    "parameters": {
      "Count": { "name": "count", "in": "query", "schema": { "type": "integer", "minimum": 1, "maximum": 100, "default": 20 } },
      "PageToken": { "name": "page_token", "in": "query", "schema": { "type": "string" } }
    },
    "responses": {
      "InvalidArgument": { "description": "Invalid request", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } },
      "Unauthenticated": { "description": "Invalid or missing bearer token", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } },
      "NotFound": { "description": "Resource not found", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } },
...
    },
    "schemas": {
      "Error": { "type": "object", "additionalProperties": false, "required": ["error"], "properties": { "error": { "type": "string" } } },
      "Pagination": {
        "type": "object",
        "required": ["count", "next_page", "previous_page", "next_page_token", "previous_page_token"],
        "properties": {
          "count": { "type": "integer" },
          "next_page": { "type": "string", "format": "uri", "nullable": true },
          "previous_page": { "type": "string", "format": "uri", "nullable": true },
          "next_page_token": { "type": "string", "nullable": true },
          "previous_page_token": { "type": "string", "nullable": true }
        }
      },
...
      "EventTypeCollection": { "type": "object", "required": ["collection", "pagination"], "properties": { "collection": { "type": "array", "items": { "$ref": "#/components/schemas/EventType" } }, "pagination": { "$ref": "#/components/schemas/Pagination" } } },
      "EventType": {
        "type": "object",
        "required": ["uri", "name", "active", "booking_method", "slug", "scheduling_url", "duration", "kind", "pooling_type", "type", "kind_description", "created_at", "updated_at", "profile", "custom_questions"],
        "properties": {
          "uri": { "type": "string", "format": "uri" }, "name": { "type": "string" }, "active": { "type": "boolean" },
          "booking_method": { "type": "string", "enum": ["instant"] }, "slug": { "type": "string" }, "scheduling_url": { "type": "string", "format": "uri" },
          "duration": { "type": "integer" }, "kind": { "type": "string", "enum": ["solo", "group"] }, "pooling_type": { "type": "string", "enum": ["collective"], "nullable": true },
          "type": { "type": "string", "enum": ["StandardEventType"] }, "kind_description": { "type": "string" },
          "created_at": { "type": "string", "format": "date-time" }, "updated_at": { "type": "string", "format": "date-time" },
          "profile": { "type": "object", "required": ["type", "name", "owner"], "properties": { "type": { "type": "string" }, "name": { "type": "string" }, "owner": { "type": "string", "format": "uri" } } },
          "custom_questions": { "type": "array", "items": { "type": "object" } },
          "color": { "type": "string", "x-not-implemented": true }, "internal_note": { "type": "string", "x-not-implemented": true }, "description_plain": { "type": "string", "x-not-implemented": true },
          "description_html": { "type": "string", "x-not-implemented": true }, "deleted_at": { "type": "string", "format": "date-time", "x-not-implemented": true },
          "secret": { "type": "boolean", "x-not-implemented": true }, "admin_managed": { "type": "boolean", "x-not-implemented": true }
        }
      },
...
    }
  }
}